requirements.txt
tensorflow
jieba
pyyaml